feat: :art: add testing env
