+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "express": "^4.17.1",
     "socket.io": "^4.1.3",
+    "socket.io-client": "^4.1.3",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
@@ -14,10 +15,12 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^3.4.0",
+    "jest": "^27.1.0",
     "prettier": "2.3.2"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node app.js",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
